requirements.txt
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.4
scikit-learn==1.3.2